+++ lib/components/Modal/botton/modal-bottom.scss
