+++ lib/components/Modal/index.ts
@@ -1,1 +1,1 @@
-export * from "./bottom";
+export * from './botton';
